Add action and app templates
