+++ jvcli/templates/2.1.0/project/actions/action.jac
@@ -1,0 +1,40 @@
+import from agent.action.{{type}} { {{node_class}} }
+
+node {{archetype}} :{{node_class}}: {
+    # Declare your has variables to be persisted here
+    # e.g has var_a : str = "string";
+
+    #* (Abilities - Uncomment and implement as needed)
+    can on_register {
+        # override to execute operations upon registration of action
+    }
+
+    can post_register {
+        # override to execute any setup code when all actions are in place
+    }
+
+    can on_enable {
+        # override to execute operations upon enabling of action
+    }
+
+    can on_disable {
+        # override to execute operations upon disabling of action
+    }
+
+    can on_deregister {
+        # override to execute operations upon deregistration of action
+    }
+
+    can touch(visitor: interact_graph_walker) -> bool {
+        # override to authorize, redirect or deny the interact walker from running execute
+    }
+
+    can execute(visitor: interact_graph_walker) -> dict {
+        # override to implement action execution
+    }
+
+    can pulse() {
+        # override to implement pulse operation
+    }
+    *#
+}
