+++ jvcli/templates/2.1.0/project/main.jac
@@ -1,2 +1,2 @@
-import:jac globals;
-include:jac jivas.agent.lib;
+import globals;
+include jivas.agent.lib;
